functions/content/process_image/requirements.txt
firebase-admin==7.1.0
google-cloud-storage==2.19.0
Pillow==11.3.0
opencv-python==4.12.0.88
